2023.8.5 - HttpResponseDto. HttpResponseInfo를 통한 에러처리 수정

+++ src/main/java/com/example/mailplugboard/controller/comment/CommentController.java
@@ -3,6 +3,8 @@ package com.example.mailplugboard.controller.comment;
 import com.example.mailplugboard.model.CommonDto;
 import com.example.mailplugboard.model.comment.CommentDto;
 import com.example.mailplugboard.model.comment.CommentListDto;
+import com.example.mailplugboard.model.httpResponse.HttpResponseDto;
+import com.example.mailplugboard.model.httpResponse.HttpResponseInfo;
 import com.example.mailplugboard.service.comment.CommentService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,13 +25,14 @@ public class CommentController {
     * 파라미터 : boardId, postId
     * */
     @GetMapping("/comment")
+    @ResponseBody
     public ResponseEntity commentsListByBoardIdAndPostId(@PathVariable("boardId") Long boardId,
                                                          @PathVariable("postId") Long postId){
         CommentListDto commentListDto = commentService.findCommentListByBoardIdAndPostId(boardId, postId);
-        if(commentListDto != null){
+        if(commentListDto.getCount() > 0){
             return new ResponseEntity(commentListDto, HttpStatus.OK);
         }else{
-            return new ResponseEntity("댓글 목록 조회 실패", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.NOT_FOUND.getStatusCode(), HttpResponseInfo.NOT_FOUND.getMessage()), HttpStatus.NOT_FOUND);
         }
     }
 
@@ -38,6 +41,7 @@ public class CommentController {
     * 파라미터 : boardId, postId, commentId
     * */
     @GetMapping("/comment/{commentId}")
+    @ResponseBody
     public ResponseEntity commentDetailByBoardIdAndPostIdAndCommentId(@PathVariable("boardId")   Long boardId,
                                                                       @PathVariable("postId")    Long postId,
                                                                       @PathVariable("commentId") Long commentId){
@@ -45,32 +49,39 @@ public class CommentController {
         if(commonDto != null){
             return new ResponseEntity(commonDto, HttpStatus.OK);
         }else{
-            return new ResponseEntity("댓글 상세 조회 실패", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.NOT_FOUND.getStatusCode(), HttpResponseInfo.NOT_FOUND.getMessage()), HttpStatus.BAD_REQUEST);
         }
     }
 
     /*
     * 특정 게시글에 댓글 등록 메소드
     * 파라미터 : commentDto(boardId, postId, displayName, password, contents)
+    * result : 1(정상), 2(등록실패), 3(해당 게시글이 없음)
     * */
     @PostMapping("/comment/write")
     @ResponseBody
     public ResponseEntity commentAddByCommentDto(@PathVariable("boardId")   Long boardId,
                                                  @PathVariable("postId")    Long postId,
                                                  @RequestBody CommentDto commentDto){
+        if(commentDto.getPassword() == null){
+            return new ResponseEntity<>(new HttpResponseDto(HttpResponseInfo.INTERNAL_SERVER_ERROR.getStatusCode(), HttpResponseInfo.INTERNAL_SERVER_ERROR.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
         commentDto.setBoardId(boardId);
         commentDto.setPostId(postId);
         int result = commentService.addCommentByCommentDto(commentDto);
-        if(result > 0){
-            return new ResponseEntity("댓글 등록 성공", HttpStatus.OK);
+        if(result == 1){
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.OK.getStatusCode(), HttpResponseInfo.OK.getMessage()), HttpStatus.OK);
+        }else if(result == 2){
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.BAD_REQUEST.getStatusCode(), HttpResponseInfo.BAD_REQUEST.getMessage()), HttpStatus.BAD_REQUEST);
         }else{
-            return new ResponseEntity("댓글 등록 실패", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.NOT_FOUND.getStatusCode(), HttpResponseInfo.NOT_FOUND.getMessage()), HttpStatus.NOT_FOUND);
         }
     }
 
     /*
     * 특정 댓글을 수정하는 메소드
     * 파라미터 : CommentDto(boardId, postId, commentId, content, password)
+    * result : 1(정상), 2(파라미터 오류), 3(해당 댓글이 없음)
     * */
     @PutMapping("/comment/{commentId}")
     @ResponseBody
@@ -81,19 +92,26 @@ public class CommentController {
         commentDto.setBoardId(boardId);
         commentDto.setPostId(postId);
         commentDto.setCommentId(commentId);
+
+        if(commentDto.getPassword() == null) {
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.INTERNAL_SERVER_ERROR.getStatusCode(), HttpResponseInfo.INTERNAL_SERVER_ERROR.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
         int result = commentService.modifyCommentByCommentDto(commentDto);
+
         if(result == 1){
-            return new ResponseEntity("댓글 수정 성공", HttpStatus.OK);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.OK.getStatusCode(), HttpResponseInfo.OK.getMessage()), HttpStatus.OK);
         }else if(result == 2){
-            return new ResponseEntity("수정 실패 : 비밀번호를 확인해 주세요", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.BAD_REQUEST.getStatusCode(), HttpResponseInfo.BAD_REQUEST.getMessage()), HttpStatus.BAD_REQUEST);
         }else{
-            return new ResponseEntity("수정 실패", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.NOT_FOUND.getStatusCode(), HttpResponseInfo.NOT_FOUND.getMessage()), HttpStatus.NOT_FOUND);
         }
     }
 
     /*
     * 댓글 삭제 메소드
     * 파라미터 CommentDto(boardId, postId, commentId, password)
+    * result : 1(정상), 2(파라미터 오류), 3(해당 댓글이 없음)
     * */
     @DeleteMapping("/comment/{commentId}")
     @ResponseBody
@@ -104,13 +122,18 @@ public class CommentController {
         commentDto.setBoardId(boardId);
         commentDto.setPostId(postId);
         commentDto.setCommentId(commentId);
+
+        if(commentDto.getPassword() == null) {
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.INTERNAL_SERVER_ERROR.getStatusCode(), HttpResponseInfo.INTERNAL_SERVER_ERROR.getMessage()), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
         int result = commentService.removeCommentByCommentDto(commentDto);
         if(result == 1){
-            return new ResponseEntity("댓글 삭제 성공", HttpStatus.OK);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.OK.getStatusCode(), HttpResponseInfo.OK.getMessage()), HttpStatus.OK);
         }else if(result == 2){
-            return new ResponseEntity("삭제 실패 : 비밀번호를 확인해 주세요", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.BAD_REQUEST.getStatusCode(), HttpResponseInfo.BAD_REQUEST.getMessage()), HttpStatus.BAD_REQUEST);
         }else{
-            return new ResponseEntity("삭제 실패", HttpStatus.BAD_REQUEST);
+            return new ResponseEntity(new HttpResponseDto(HttpResponseInfo.NOT_FOUND.getStatusCode(), HttpResponseInfo.NOT_FOUND.getMessage()), HttpStatus.NOT_FOUND);
         }
     }
